requirements.txt
apache-airflow==2.10.2
pandas
numpy
sqlalchemy
psycopg2-binary
matplotlib
pytest
